src/core/BaseKalendar.ts
// src/core/BaseCalendar.ts

import { CalendarSystem, CalendarDate, CalendarEvent, CalendarTheme, DatePattern, MonthCustomization } from './CalendarSystem';
import { CalendarDayView, CalendarWeekView, CalendarMonthView, CalendarYearView } from './CalendarViews';

export abstract class BaseKalendar implements CalendarSystem {
  protected events: CalendarEvent[] = [];
  protected themes: CalendarTheme[] = [];
  protected monthCustomizations: Map<number, MonthCustomization> = new Map();
  protected dayIntentions: Map<string, string> = new Map();
  protected weekIntentions: Map<string, string> = new Map();
  protected monthIntentions: Map<string, string> = new Map();

  abstract get name(): string;
  abstract isLeapYear(year: number): boolean;
  abstract daysInYear(year: number): number;
  abstract daysInMonth(year: number, month: number): number;
  abstract dateToOrdinal(date: CalendarDate): number;
  abstract ordinalToDate(year: number, ordinal: number): CalendarDate;
  abstract addDays(date: CalendarDate, days: number): CalendarDate;
  abstract dayOfWeek(date: CalendarDate): number;

  protected dateMatches(date: CalendarDate, pattern: DatePattern): boolean {
    switch (pattern.type) {
      case 'specific':
        return date.year === pattern.date.year && 
               date.month === pattern.date.month && 
               date.day === pattern.date.day;
      case 'weekly':
        return this.dayOfWeek(date) === pattern.dayOfWeek;
      case 'monthly':
        return date.day === pattern.dayOfMonth;
      case 'yearly':
        return date.month === pattern.month && date.day === pattern.day;
      case 'custom':
        return pattern.predicate(date);
    }
  }

  formatDate(date: CalendarDate, format: string): string {
    // Implement date formatting logic
    return `${date.year}-${String(date.month).padStart(2, '0')}-${String(date.day).padStart(2, '0')}`;
  }

  parseDate(dateString: string, format: string): CalendarDate {
    // Implement date parsing logic
    const [year, month, day] = dateString.split('-').map(Number);
    return { year, month, day };
  }

  addEvent(event: CalendarEvent): void {
    this.events.push(event);
  }

  getEvents(date: CalendarDate): CalendarEvent[] {
    return this.events.filter(event => this.dateMatches(date, event.pattern));
  }

  addTheme(theme: CalendarTheme): void {
    this.themes.push(theme);
  }

  getTheme(date: CalendarDate): CalendarTheme | null {
    return this.themes.find(theme => this.dateMatches(date, theme.pattern)) || null;
  }

  customizeMonth(month: number, customization: MonthCustomization): void {
    this.monthCustomizations.set(month, customization);
  }

  getMonthCustomization(month: number): MonthCustomization | null {
    return this.monthCustomizations.get(month) || null;
  }

  setDayIntention(date: CalendarDate, intention: string): void {
    const key = `${date.year}-${date.month}-${date.day}`;
    this.dayIntentions.set(key, intention);
  }

  getDayIntention(date: CalendarDate): string | null {
    const key = `${date.year}-${date.month}-${date.day}`;
    return this.dayIntentions.get(key) || null;
  }

  setWeekIntention(year: number, weekNumber: number, intention: string): void {
    const key = `${year}-${weekNumber}`;
    this.weekIntentions.set(key, intention);
  }

  getWeekIntention(year: number, weekNumber: number): string | null {
    const key = `${year}-${weekNumber}`;
    return this.weekIntentions.get(key) || null;
  }

  setMonthIntention(year: number, month: number, intention: string): void {
    const key = `${year}-${month}`;
    this.monthIntentions.set(key, intention);
  }

  getMonthIntention(year: number, month: number): string | null {
    const key = `${year}-${month}`;
    return this.monthIntentions.get(key) || null;
  }

  generateDayView(date: CalendarDate): CalendarDayView {
    return {
      date,
      events: this.getEvents(date),
      theme: this.getTheme(date),
      intention: this.getDayIntention(date)
    };
  }

  generateWeekView(startDate: CalendarDate): CalendarWeekView {
    const days: CalendarDayView[] = [];
    for (let i = 0; i < 7; i++) {
      const date = this.addDays(startDate, i);
      days.push(this.generateDayView(date));
    }
    return {
      startDate,
      endDate: this.addDays(startDate, 6),
      days,
      intention: this.getWeekIntention(startDate.year, this.getWeekNumber(startDate))
    };
  }

  generateMonthView(year: number, month: number): CalendarMonthView {
    const firstDay = { year, month, day: 1 };
    const daysInMonth = this.daysInMonth(year, month);
    const weeks: CalendarWeekView[] = [];
    
    let currentDay = firstDay;
    
    while (currentDay.day <= daysInMonth) {
      weeks.push(this.generateWeekView(currentDay));
      currentDay = this.addDays(currentDay, 7);
    }

    return {
      year,
      month,
      weeks,
      customization: this.getMonthCustomization(month),
      intention: this.getMonthIntention(year, month)
    };
  }

  getMonthRange(): { min: number; max: number } {
    // Implement the getMonthRange method here
    return { min: 1, max: 12 };
  }

  generateYearView(year: number): CalendarYearView {
    const months: CalendarMonthView[] = [];
    const { min, max } = this.getMonthRange();
    for (let month = min; month <= max; month++) {
      months.push(this.generateMonthView(year, month));
    }
    return { year, months };
  }

  protected getWeekNumber(date: CalendarDate): number {
    // This is a simplified implementation. You might want to adjust this based on your specific calendar rules.
    return Math.ceil(this.dateToOrdinal(date) / 7);
  }
}